refactor: split oversized files into part files to reduce LOC
Normalize lib/services imports to package: paths (prep), then break the
largest standalone files into part files (verbatim moves, no behavior change):

- web_market_data_service.dart -> web_market_data_results.dart (result types)
- portfolio_rebalance_service.dart -> portfolio_rebalance_models.dart (DTOs/enums)
- pdf_table_reconstructor.dart -> pdf_table_models.dart (internal types)
- pillar_detail_screen.dart -> pillar_detail_overview.dart + pillar_detail_cards.dart
- ath_celebration_overlay.dart -> ath_overlay_body.dart + ath_fireworks.dart

No public API changes; imports and tests unchanged.
