+++ lib/services/buffer_service.dart
@@ -1,7 +1,7 @@
 import 'package:drift/drift.dart';
 
-import '../database/database.dart';
-import '../utils/logger.dart';
+import 'package:finance_copilot/database/database.dart';
+import 'package:finance_copilot/utils/logger.dart';
 
 final _log = getLogger('BufferService');
