+++ lib/services/income_service.dart
@@ -1,8 +1,8 @@
 import 'package:drift/drift.dart';
 
-import '../database/database.dart';
-import '../database/tables.dart';
-import '../utils/logger.dart';
+import 'package:finance_copilot/database/database.dart';
+import 'package:finance_copilot/database/tables.dart';
+import 'package:finance_copilot/utils/logger.dart';
 
 final _log = getLogger('IncomeService');
