+++ lib/services/transaction_service.dart
@@ -2,10 +2,10 @@ import 'dart:convert';
 
 import 'package:drift/drift.dart';
 
-import '../database/database.dart';
-import '../database/tables.dart';
-import '../utils/amount_parser.dart' as amt;
-import '../utils/logger.dart';
+import 'package:finance_copilot/database/database.dart';
+import 'package:finance_copilot/database/tables.dart';
+import 'package:finance_copilot/utils/amount_parser.dart' as amt;
+import 'package:finance_copilot/utils/logger.dart';
 
 final _log = getLogger('TransactionService');
